lab4/lab_4_back/lab4/beer/serializer.py
from rest_framework import serializers
from .models import Company

class MySerializer(serializers.ModelSerializer):
    id = serializers.IntegerField(required=False)

    class Meta:
        model = Company
        fields = '__all__'