lab4/lab_4_back/lab4/requirements.txt
asgiref==3.3.4
Django==3.2
django-rest-framework==0.1.0
djangorestframework==3.12.2
django-cors-headers==3.7.0
redis==3.5.3
